requirements.txt
fastapi
sqlalchemy
passlib
PyJWT
pydantic[email]
mysqlclient
python-multipart
bcrypt
uvicorn
requests